Extract `rgh-deduplicator`
