+++ source/feature-manager.tsx
@@ -1,5 +1,4 @@
 import React from 'dom-chef';
-import {$optional} from 'select-dom/strict.js';
 import {elementExists} from 'select-dom';
 import domLoaded from 'dom-loaded';
 import stripIndent from 'strip-indent';
@@ -41,7 +40,7 @@ type FeatureLoader = RunConditions & {
 	awaitDomReady?: true;
 
 	/**
-	When pressing the back button, DOM changes and listeners are still there. Using a selector here would use the integrated deduplication logic, but it cannot be used with `delegate` and it shouldn't use `has-rgh` and `has-rgh-inner` anymore. #5871 #
+	When pressing the back button, DOM changes and listeners are still there. Using a selector here would use the integrated deduplication logic, but it cannot be used with `delegate` and it shouldn't use `has-rgh` and `has-rgh-inner` anymore. #5871
 	@deprecated
 	@default false
 	*/
@@ -215,23 +214,6 @@ function unloadAll(): void {
 
 	currentFeatureControllers.clear();
 }
-
-/*
-When navigating back and forth in history, GitHub will preserve the DOM changes;
-This means that the old features will still be on the page and don't need to re-run.
-
-This marks each as "processed"
-*/
-void add('rgh-deduplicator' as FeatureID, {
-	awaitDomReady: true,
-	async init() {
-		// `await` kicks it to the next tick, after the other features have checked for 'has-rgh', so they can run once.
-		await Promise.resolve();
-		$optional('has-rgh')?.remove(); // https://github.com/refined-github/refined-github/issues/6568
-		$optional(_`#js-repo-pjax-container, #js-pjax-container`)?.append(<has-rgh />);
-		$optional(_`turbo-frame`)?.append(<has-rgh-inner />); // #4567
-	},
-});
 
 const features = {
 	add,
